requirements.txt
flask
biopython
xmltodict
